- Added storage util projects - Extended database validation with exception to log - Replaced *DatabaseStructure.sql* with an equivalent version, generated by Enterprise Architect Related to Issue [WTI-434]
git-svn-id: https://repos.deltares.nl/repos/WettelijkToetsInstrumentarium/trunk@2503 5a3db67b-de53-47b5-99c8-a1c30a6650e2

Former-commit-id: 45035258828b9ec49ab364b8f2ced9d549d50e73
Former-commit-id: 48ced0cc6de6dcbbd28d18cadac5ddb909c82384

+++ Application/Ringtoets/src/Application.Ringtoets.Storage/DbContext/ProjectEntity.cs
@@ -44,7 +44,7 @@ namespace Application.Ringtoets.Storage.DbContext
     
         public long ProjectEntityId { get; set; }
         public string Description { get; set; }
-        public Nullable<long> LastUpdated { get; set; }
+        public Nullable<decimal> LastUpdated { get; set; }
     
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2227:CollectionPropertiesShouldBeReadOnly")]
         public virtual ICollection<DikeAssessmentSectionEntity> DikeAssessmentSectionEntities { get; set; }
